backend/generate_audio_live/requirements.txt
functions-framework
google-genai
flask
